pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cli-encrypt-txt"
version = "3.1.0"
description = "Hyper-secure Quantum-Ready CLI file encryptor: AES-256-GCM + Argon2id with opaque steganographic headers"
authors = [
  { name="Adityakrishna LDK SreeRamachandrarao", email="adityadronamraju20@gmail.com"}
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
  "cryptography>=42.0.0",
  "argon2-cffi>=23.1.0"
]

[project.urls]
Homepage = "https://github.com/sreeramachand/cli-encrypt-txt"
Source = "https://github.com/sreeramachand/cli-encrypt-txt"

[project.scripts]
cli-encrypt-txt = "cli_encrypt_txt.__main__:main"
